Merfe into single file
